Adds optional list of domains to accept
This adds a new preference key that will either allow any or a whitelist of domains to be appended to the username value even when the ADDomain value is managed. A boolean true will allow any domain to login by specifying to entire UPN. An array value will provide a whitelist of domains to allow. No value continues the previous bavahior of only allowing the managed domain to login.