Adds creating admin accounts from AD group
This adds a new preference key to allow promoting a newly created local account to a local admin by its membership in one or more AD groups. Pref key should be an array of group names. Membership in at least one of these groups will flag the account to be created as an admin